chore: move landing page into group
