+++ src/app/globals.css
@@ -1,7 +1,7 @@
 @import 'tailwindcss';
 
 :root {
-  --background: #766579;
+  --background: #796d65;
   --foreground: #ededed;
 }
